Various bug fixes and improvements

+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/openWeatherMap/common/IOWMCloudsResponse.kt
@@ -4,6 +4,6 @@ package ru.mts.avpopo85.weathery.data.model.base.openWeatherMap.common
 interface IOWMCloudsResponse {
 
     /**Cloudiness, %*/
-    val cloudiness: Int
+    val cloudiness: Double
 
 }
